pyproject.toml
[project]
name = "mockoidc"
version = "1.0.0"
description = "Mock OpenID Connect Provider for testing and development"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Marco", email = "marco@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-jose[cryptography]==3.3.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "cryptography==42.0.0",
    "python-multipart==0.0.6",
    "flask==3.0.0",
    "requests==2.31.0",
    "httpx==0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx==0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "main.py",
    "config.py",
    "models.py",
    "token_service.py",
    "jwks_service.py",
    "claims_generator.py",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []
